Add block_device_common
