+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod block_device;
+pub mod block_device_common;
 pub mod config;
 pub mod device_manager;
 pub mod grpc_server;
